refactor(commands): extract account/auth commands into accounts.rs
